// frontend/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BookResponse {
  archived?: boolean;
  author?: string;
  cover?: Array<string>;
  id?: number;
  isbn?: string;
  ownerName?: string;
  publisher?: string;
  rate?: number;
  ratedBy?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
}